+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM python:3.9
+FROM python:3.8-slim-buster
 
 ENV PYTHONUNBUFFERED=1
